refactor: test classes and added testprovider
