+++ src/auth/Components/Validator.js
@@ -1,12 +1,12 @@
 import React from 'react';
-import { useGet } from '../hooks/useRequest';
+import { useGet } from '../../hooks/useRequest';
 
-import Constants from '../util/Constants';
+import Constants from '../../util/Constants';
 
 const ComponentValidator = ({ allowedAuthorities, authorityKey, children }) => {
   const { data, status } = useGet(Constants.REACT_QUERY.KEYS.GET_USER_LOGGED, 'GET_USER_LOGGED', null);
 
-  const getRoute = () => {
+  const getComponentValidated = () => {
     if (status === Constants.REACT_QUERY.CODES.SUCCESS) {
       const hasAccess = data.data.user[authorityKey].some(permission => allowedAuthorities.includes(permission));
       if (hasAccess) {
@@ -17,7 +17,7 @@ const ComponentValidator = ({ allowedAuthorities, authorityKey, children }) => {
     return <></>;
   };
 
-  return getRoute();
+  return getComponentValidated();
 };
 
 export default ComponentValidator;
