+++ src/pages/Todo/List/TodoList.js
@@ -4,7 +4,7 @@ import { Helmet } from 'react-helmet';
 
 import CenteredLink from '../../../components/Link/Link';
 import BrowserURL from '../../../util/BrowserURL';
-import ComponentValidator from '../../../auth/ComponentValidator';
+import ComponentValidator from '../../../auth/Components/Validator';
 
 export default function TodoList() {
   const history = useHistory();
